# gestion-de-stock-frontend-master/tools/swagger/src/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger Api documentation for Gestion de stock application",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "Swagger Api Documentation for Gestion de stock application",
    "contact" : {
      "name" : "Gestion de Stock Team",
      "url" : "",
      "email" : ""
    }
  },
  "basePath" : "http://localhost:8081",
...
  "paths" : {
...
    "/gestiondestock/v1/articles/{idArticle}" : {
      "get" : {
        "tags" : [ "articles" ],
        "summary" : "Rechercher un article par ID",
        "description" : "Cette methode permet de chercher un article par son ID",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idArticle",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "L'article a ete trouve dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/ArticleDto"
            }
          },
          "404" : {
            "description" : "Aucun article n'existe dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/gestiondestock/v1/categories/{idCategory}" : {
      "get" : {
        "tags" : [ "categories" ],
        "summary" : "Rechercher une categorie par ID",
        "description" : "Cette methode permet de chercher une categorie par son ID",
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idCategory",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "La categorie a ete trouve dans la BDD",
            "schema" : {
              "$ref" : "#/definitions/CategoryDto"
            }
          },
          "404" : {
            "description" : "Aucune categorie n'existe dans la BDD avec l'ID fourni"
          }
        }
      }
    },
...
    "/gestiondestock/v1/clients/{idClient}" : {
      "get" : {
        "tags" : [ "clients" ],
        "operationId" : "findById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idClient",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ClientDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/commandesclients/{idCommandeClient}" : {
      "get" : {
        "tags" : [ "commandesclients" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idCommandeClient",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeClientDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/commandesfournisseurs/update/fournisseur/{idCommande}/{idFournisseur}" : {
      "patch" : {
        "tags" : [ "commandefournisseur" ],
        "operationId" : "updateFournisseur",
        "parameters" : [ {
          "name" : "idCommande",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "idFournisseur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeFournisseurDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/commandesfournisseurs/{idCommandeFournisseur}" : {
      "get" : {
        "tags" : [ "commandefournisseur" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idCommandeFournisseur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CommandeFournisseurDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/entreprises/{idEntreprise}" : {
      "get" : {
        "tags" : [ "entreprises" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idEntreprise",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/EntrepriseDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/fournisseurs/create" : {
      "post" : {
        "tags" : [ "fournisseur" ],
        "operationId" : "save",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/FournisseurDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FournisseurDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/fournisseurs/delete/{idFournisseur}" : {
      "delete" : {
        "tags" : [ "fournisseur" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "idFournisseur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/gestiondestock/v1/fournisseurs/{idFournisseur}" : {
      "get" : {
        "tags" : [ "fournisseur" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idFournisseur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FournisseurDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/mvtstk/correctionneg" : {
      "post" : {
        "tags" : [ "mvtstk" ],
        "operationId" : "correctionStockNeg",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/MvtStkDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MvtStkDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/mvtstk/correctionpos" : {
      "post" : {
        "tags" : [ "mvtstk" ],
        "operationId" : "correctionStockPos",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/MvtStkDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MvtStkDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/mvtstk/entree" : {
      "post" : {
        "tags" : [ "mvtstk" ],
        "operationId" : "entreeStock",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/MvtStkDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MvtStkDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/mvtstk/filter/article/{idArticle}" : {
      "get" : {
        "tags" : [ "mvtstk" ],
        "operationId" : "mvtStkArticle",
        "parameters" : [ {
          "name" : "idArticle",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/MvtStkDto"
              }
            }
          }
        }
      }
    },
    "/gestiondestock/v1/mvtstk/sortie" : {
      "post" : {
        "tags" : [ "mvtstk" ],
        "operationId" : "sortieStock",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/MvtStkDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MvtStkDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/mvtstk/stockreel/{idArticle}" : {
      "get" : {
        "tags" : [ "mvtstk" ],
        "operationId" : "stockReelArticle",
        "parameters" : [ {
          "name" : "idArticle",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "number"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/save/{id}/{title}/{context}" : {
      "post" : {
        "tags" : [ "photos" ],
        "operationId" : "savePhoto",
        "parameters" : [ {
          "name" : "context",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "file",
          "in" : "formData",
          "required" : true,
          "type" : "file"
        }, {
          "name" : "title",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/utilisateurs/create" : {
      "post" : {
        "tags" : [ "utilisateurs" ],
        "operationId" : "save",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UtilisateurDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UtilisateurDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/utilisateurs/delete/{idUtilisateur}" : {
      "delete" : {
        "tags" : [ "utilisateurs" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "idUtilisateur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/gestiondestock/v1/utilisateurs/find/{email}" : {
      "get" : {
        "tags" : [ "utilisateurs" ],
        "operationId" : "findByEmail",
        "parameters" : [ {
          "name" : "email",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UtilisateurDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/utilisateurs/update/password" : {
      "post" : {
        "tags" : [ "utilisateurs" ],
        "operationId" : "changerMotDePasse",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ChangerMotDePasseUtilisateurDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UtilisateurDto"
            }
          }
        }
      }
    },
    "/gestiondestock/v1/utilisateurs/{idUtilisateur}" : {
      "get" : {
        "tags" : [ "utilisateurs" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idUtilisateur",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UtilisateurDto"
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/ventes/{idVente}" : {
      "get" : {
        "tags" : [ "ventes" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "idVente",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/VentesDto"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Adresse" : {
      "type" : "object",
      "properties" : {
        "adresse1" : {
          "type" : "string"
        },
        "adresse2" : {
          "type" : "string"
        },
        "ville" : {
          "type" : "string"
        },
        "codePostale" : {
          "type" : "string"
        },
        "pays" : {
          "type" : "string"
        }
      }
    },
    "AdresseDto" : {
      "type" : "object",
      "properties" : {
        "adresse1" : {
          "type" : "string"
        },
        "adresse2" : {
          "type" : "string"
        },
        "ville" : {
          "type" : "string"
        },
        "codePostale" : {
          "type" : "string"
        },
        "pays" : {
          "type" : "string"
        }
      }
    },
    "Article" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "codeArticle" : {
          "type" : "string"
        },
        "designation" : {
          "type" : "string"
        },
        "prixUnitaireHt" : {
          "type" : "number"
        },
        "tauxTva" : {
          "type" : "number"
        },
        "prixUnitaireTtc" : {
          "type" : "number"
        },
        "photo" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "category" : {
          "$ref" : "#/definitions/Category"
        },
        "ligneVentes" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneVente"
          }
        },
        "ligneCommandeClients" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeClient"
          }
        },
        "ligneCommandeFournisseurs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeFournisseur"
          }
        },
        "mvtStks" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/MvtStk"
          }
        }
      }
    },
    "ArticleDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "codeArticle" : {
          "type" : "string"
        },
        "designation" : {
          "type" : "string"
        },
        "prixUnitaireHt" : {
          "type" : "number"
        },
        "tauxTva" : {
          "type" : "number"
        },
        "prixUnitaireTtc" : {
          "type" : "number"
        },
        "photo" : {
          "type" : "string"
        },
        "category" : {
          "$ref" : "#/definitions/CategoryDto"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "Category" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "code" : {
          "type" : "string"
        },
        "designation" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "articles" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Article"
          }
        }
      }
    },
    "CategoryDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "code" : {
          "type" : "string"
        },
        "designation" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "ChangerMotDePasseUtilisateurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "motDePasse" : {
          "type" : "string"
        },
        "confirmMotDePasse" : {
          "type" : "string"
        }
      }
    },
...
    "ClientDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/AdresseDto"
        },
        "photo" : {
          "type" : "string"
        },
        "mail" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "CommandeClientDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "code" : {
          "type" : "string"
        },
        "dateCommande" : {
          "type" : "integer",
          "format" : "int64"
        },
        "etatCommande" : {
          "type" : "string",
          "enum" : [ "EN_PREPARATION", "VALIDEE", "LIVREE" ]
        },
        "client" : {
          "$ref" : "#/definitions/ClientDto"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "ligneCommandeClients" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeClientDto"
          }
        },
        "commandeLivree" : {
          "type" : "boolean"
        }
      }
    },
    "CommandeFournisseur" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "code" : {
          "type" : "string"
        },
        "dateCommande" : {
          "type" : "integer",
          "format" : "int64"
        },
        "etatCommande" : {
          "type" : "string",
          "enum" : [ "EN_PREPARATION", "VALIDEE", "LIVREE" ]
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "fournisseur" : {
          "$ref" : "#/definitions/Fournisseur"
        },
        "ligneCommandeFournisseurs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeFournisseur"
          }
        }
      }
    },
    "CommandeFournisseurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "code" : {
          "type" : "string"
        },
        "dateCommande" : {
          "type" : "integer",
          "format" : "int64"
        },
        "etatCommande" : {
          "type" : "string",
          "enum" : [ "EN_PREPARATION", "VALIDEE", "LIVREE" ]
        },
        "fournisseur" : {
          "$ref" : "#/definitions/FournisseurDto"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "ligneCommandeFournisseurs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneCommandeFournisseurDto"
          }
        },
        "commandeLivree" : {
          "type" : "boolean"
        }
      }
    },
    "EntrepriseDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/AdresseDto"
        },
        "codeFiscal" : {
          "type" : "string"
        },
        "photo" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "steWeb" : {
          "type" : "string"
        }
      }
    },
    "Fournisseur" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/Adresse"
        },
        "photo" : {
          "type" : "string"
        },
        "mail" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        },
        "commandeFournisseurs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CommandeFournisseur"
          }
        }
      }
    },
    "FournisseurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/AdresseDto"
        },
        "photo" : {
          "type" : "string"
        },
        "mail" : {
          "type" : "string"
        },
        "numTel" : {
          "type" : "string"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneCommandeClient" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "article" : {
          "$ref" : "#/definitions/Article"
        },
        "commandeClient" : {
          "$ref" : "#/definitions/CommandeClient"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneCommandeClientDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDto"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneCommandeFournisseur" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "article" : {
          "$ref" : "#/definitions/Article"
        },
        "commandeFournisseur" : {
          "$ref" : "#/definitions/CommandeFournisseur"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneCommandeFournisseurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDto"
        },
        "commandeFournisseur" : {
          "$ref" : "#/definitions/CommandeFournisseur"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneVente" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastModifiedDate" : {
          "type" : "integer",
          "format" : "int64"
        },
        "vente" : {
          "$ref" : "#/definitions/Ventes"
        },
        "article" : {
          "$ref" : "#/definitions/Article"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "LigneVenteDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "vente" : {
          "$ref" : "#/definitions/VentesDto"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDto"
        },
        "quantite" : {
          "type" : "number"
        },
        "prixUnitaire" : {
          "type" : "number"
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "MvtStkDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "dateMvt" : {
          "type" : "integer",
          "format" : "int64"
        },
        "quantite" : {
          "type" : "number"
        },
        "article" : {
          "$ref" : "#/definitions/ArticleDto"
        },
        "typeMvt" : {
          "type" : "string",
          "enum" : [ "ENTREE", "SORTIE", "CORRECTION_POS", "CORRECTION_NEG" ]
        },
        "sourceMvt" : {
          "type" : "string",
          "enum" : [ "COMMANDE_CLIENT", "COMMANDE_FOURNISSEUR", "VENTE" ]
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "RolesDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "roleName" : {
          "type" : "string"
        }
      }
    },
    "UtilisateurDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nom" : {
          "type" : "string"
        },
        "prenom" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "dateDeNaissance" : {
          "type" : "integer",
          "format" : "int64"
        },
        "moteDePasse" : {
          "type" : "string"
        },
        "adresse" : {
          "$ref" : "#/definitions/AdresseDto"
        },
        "photo" : {
          "type" : "string"
        },
        "entreprise" : {
          "$ref" : "#/definitions/EntrepriseDto"
        },
        "roles" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/RolesDto"
          }
        }
      }
    },
...
    "VentesDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "code" : {
          "type" : "string"
        },
        "dateVente" : {
          "type" : "integer",
          "format" : "int64"
        },
        "commentaire" : {
          "type" : "string"
        },
        "ligneVentes" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LigneVenteDto"
          }
        },
        "idEntreprise" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    }
  }
}